[compat, modify] Use numeric_limits insted of LONG_MAX.

+++ src/lib/coil/common/coil/Properties.cpp
@@ -71,7 +71,7 @@ namespace coil
    * @brief Constructor(Give the default value with char*[])
    * @endif
    */
-  Properties::Properties(const char* const defaults[], long num)
+  Properties::Properties(const char* const defaults[], size_t num)
   {
     leaf.clear();
     setDefaults(defaults, num);
@@ -306,9 +306,9 @@ namespace coil
    * @brief Set a default value together in the property list
    * @endif
    */
-  void Properties::setDefaults(const char* const defaults[], long num)
+  void Properties::setDefaults(const char* const defaults[], size_t num)
   {
-    for (long i = 0; i < num && defaults[i][0] != '\0' ; i += 2)
+    for (size_t i = 0; i < num && defaults[i][0] != '\0' ; i += 2)
       {
         setDefault(eraseBothEndsBlank(defaults[i]),
                    eraseBothEndsBlank(defaults[i + 1]));
